Separate Port view models and views
