+++ src/DynamoCoreWpf/ViewModels/Core/PortViewModel.cs
@@ -14,30 +14,18 @@ namespace Dynamo.ViewModels
     {
         #region Properties/Fields
 
-        private readonly PortModel _port;
-        private readonly NodeViewModel _node;
+        protected readonly PortModel _port;
+        protected readonly NodeViewModel _node;
         private DelegateCommand _useLevelsCommand;
         private DelegateCommand _keepListStructureCommand;
-        private DelegateCommand _breakConnectionsCommand;
-        private DelegateCommand _hideConnectionsCommand;
-        private const double autocompletePopupSpacing = 2.5;
-        private SolidColorBrush portBorderBrushColor = new SolidColorBrush(Color.FromArgb(255, 204, 204, 204));
-        private SolidColorBrush portValueMarkerColor = new SolidColorBrush(Color.FromArgb(255, 204, 204, 204));
-        private SolidColorBrush portBackgroundColor = new SolidColorBrush(Color.FromArgb(0, 60, 60, 60));
-        internal bool inputPortDisconnectedByConnectCommand = false;
         private bool _showUseLevelMenu;
-        private bool areConnectorsHidden;
-        private string showHideWiresButtonContent = "";
-        private bool hideWiresButtonEnabled;
+        private const double autocompletePopupSpacing = 2.5;
+        internal bool inputPortDisconnectedByConnectCommand = false;
+        protected static SolidColorBrush portBorderBrushColor = new SolidColorBrush(Color.FromArgb(255, 204, 204, 204));
+        protected static SolidColorBrush portBackgroundColor = new SolidColorBrush(Color.FromArgb(0, 60, 60, 60));
 
-        public static SolidColorBrush PortValueMarkerBlue = new SolidColorBrush(Color.FromRgb(106, 192, 231));
-        public static SolidColorBrush PortValueMarkerRed = new SolidColorBrush(Color.FromRgb(235, 85, 85));
-
-        public static SolidColorBrush PortBackgroundColorDefault = new SolidColorBrush(Color.FromRgb(60, 60, 60));
-        public static SolidColorBrush PortBackgroundColorKeepListStructure = new SolidColorBrush(Color.FromRgb(83, 126, 145));
-
-        public static SolidColorBrush PortBorderBrushColorDefault = new SolidColorBrush(Color.FromRgb(161, 161, 161));
-        public static SolidColorBrush PortBorderBrushColorKeepListStructure = new SolidColorBrush(Color.FromRgb(168, 181, 187));
+        protected static readonly SolidColorBrush PortBackgroundColorDefault = new SolidColorBrush(Color.FromRgb(60, 60, 60));
+        protected static readonly SolidColorBrush PortBorderBrushColorDefault = new SolidColorBrush(Color.FromRgb(161, 161, 161));
 
         /// <summary>
         /// Port model.
@@ -135,6 +123,7 @@ namespace Dynamo.ViewModels
         /// <summary>
         /// Returns whether this port has a default value that can be used.
         /// </summary>
+        [Obsolete("This method will be removed in Dynamo 3.0 - please use the InPortViewModel")]
         public bool DefaultValueEnabled
         {
             get { return _port.DefaultValue != null; }
@@ -143,6 +132,7 @@ namespace Dynamo.ViewModels
         /// <summary>
         /// Returns whether the port is using its default value, or whether this been disabled
         /// </summary>
+        [Obsolete("This method will be removed in Dynamo 3.0 - please use the InPortViewModel")]
         public bool UsingDefaultValue
         {
             get { return _port.UsingDefaultValue; }
@@ -177,6 +167,7 @@ namespace Dynamo.ViewModels
         /// <summary>
         /// If should display Use Levels popup menu. 
         /// </summary>
+        [Obsolete("This method will be removed in Dynamo 3.0 - please use the InPortViewModel")]
         public bool ShowUseLevelMenu
         {
             get
@@ -193,27 +184,16 @@ namespace Dynamo.ViewModels
         /// <summary>
         /// If UseLevel is enabled on this port.
         /// </summary>
+        [Obsolete("This method will be removed in Dynamo 3.0 - please use the InPortViewModel")]
         public bool UseLevels
         {
             get { return _port.UseLevels; }
         }
 
         /// <summary>
-        /// Determines whether or not the UseLevelsSpinner is visible on the port.
-        /// </summary>
-        public Visibility UseLevelSpinnerVisible
-        {
-            get
-            {
-                if (PortType == PortType.Output) return Visibility.Collapsed;
-                if (UseLevels) return Visibility.Visible;
-                return Visibility.Hidden;
-            }
-        }
-
-        /// <summary>
         /// If should keep list structure on this port.
         /// </summary>
+        [Obsolete("This method will be removed in Dynamo 3.0 - please use the InPortViewModel")]
         public bool ShouldKeepListStructure
         {
             get { return _port.KeepListStructure; }
@@ -255,98 +235,6 @@ namespace Dynamo.ViewModels
         }
         
         /// <summary>
-        /// Shows or hides the Use Levels and Keep List Structure checkboxes
-        /// in the node chevron popup menu.
-        /// </summary>
-        public bool UseLevelCheckBoxVisibility
-        {
-            get => _port.PortType == PortType.Input;
-        }
-
-        /// <summary>
-        /// Shows or hides the Use Default Value checkbox in the node chevron popup menu.
-        /// </summary>
-        public bool UseDefaultValueCheckBoxVisibility
-        {
-            get => _port.PortType == PortType.Input && DefaultValueEnabled;
-        }
-
-        /// <summary>
-        /// Shows or hides the Break Connections, Hide Wires and UnhideWires buttons in the node chevron popup menu.
-        /// </summary>
-        public bool OutputPortConnectionsButtonsVisible
-        {
-            get => _port.PortType == PortType.Output;
-        }
-
-        /// <summary>
-        /// Enables or disables the Break Connections button on the node output port context menu.
-        /// </summary>
-        public bool OutputPortBreakConnectionsButtonEnabled
-        {
-            get => OutputPortConnectionsButtonsVisible && IsConnected;
-        }
-
-        /// <summary>
-        /// Determines whether the output port button says 'Hide Wires' or 'Show Wires'
-        /// </summary>
-        public string ShowHideWiresButtonContent
-        {
-            get => showHideWiresButtonContent;
-            set
-            {
-                showHideWiresButtonContent = value;
-                RaisePropertyChanged(nameof(ShowHideWiresButtonContent));
-            }
-        }
-
-        /// <summary>
-        /// Sets the visibility of the connectors from the port. This will overwrite the 
-        /// individual visibility of the connectors. However when visibility is controlled 
-        /// from the connector, that connector's visibility will overwrite its previous state.
-        /// In order to overwrite visibility of all connectors associated with a port, us this 
-        /// flag again.
-        /// </summary>
-        public bool SetConnectorsVisibility
-        {
-            get => areConnectorsHidden;
-            set
-            {
-                areConnectorsHidden = value; 
-                RaisePropertyChanged(nameof(SetConnectorsVisibility));
-            }
-        }
-
-        /// <summary>
-        /// Enables or disables the Hide Wires button on the node output port context menu.
-        /// </summary>
-        public bool HideWiresButtonEnabled
-        {
-            get => hideWiresButtonEnabled;
-            set
-            {
-                hideWiresButtonEnabled = value; 
-                RaisePropertyChanged(nameof(HideWiresButtonEnabled));
-            }
-        }
-
-        /// <summary>
-        /// Takes care of the multiple UI concerns when dealing with the Unhide/Hide Wires button
-        /// on the output port's context menu.
-        /// </summary>
-        private void RefreshHideWiresButton()
-        {
-            HideWiresButtonEnabled = _port.Connectors.Count > 0;
-            SetConnectorsVisibility = CheckIfConnectorsAreHidden();
-
-            ShowHideWiresButtonContent = SetConnectorsVisibility
-                ? Properties.Resources.UnhideWiresPopupMenuItem
-                : Properties.Resources.HideWiresPopupMenuItem;
-
-            RaisePropertyChanged(nameof(ShowHideWiresButtonContent));
-        }
-
-        /// <summary>
         /// Sets the color of the port's border brush
         /// </summary>
         public SolidColorBrush PortBorderBrushColor
@@ -356,19 +244,6 @@ namespace Dynamo.ViewModels
             {
                 portBorderBrushColor = value;
                 RaisePropertyChanged(nameof(PortBorderBrushColor));
-            }
-        }
-
-        /// <summary>
-        /// Sets the color of the small rectangular marker on each input port.
-        /// </summary>
-        public SolidColorBrush PortValueMarkerColor
-        {
-            get => portValueMarkerColor;
-            set
-            {
-                portValueMarkerColor = value;
-                RaisePropertyChanged(nameof(PortValueMarkerColor));
             }
         }
 
@@ -383,21 +258,6 @@ namespace Dynamo.ViewModels
             {
                 portBackgroundColor = value;
                 RaisePropertyChanged(nameof(PortBackgroundColor));
-            }
-        }
-
-        /// <summary>
-        /// Sets the color of the use levels popup in the input port context menu.
-        /// This changes when the Keep List Structure option is activated and the port
-        /// is connected, upon which it turns blue.
-        /// </summary>
-        public SolidColorBrush UseLevelsMenuColor
-        {
-            get
-            {
-                return ShouldKeepListStructure && _port.IsConnected
-                    ? new SolidColorBrush(Color.FromArgb(255, 60, 60, 60))
-                    : new SolidColorBrush(Color.FromArgb(255, 83, 83, 83));
             }
         }
 
@@ -420,7 +280,6 @@ namespace Dynamo.ViewModels
             _node.WorkspaceViewModel.PropertyChanged += Workspace_PropertyChanged;
 
             RefreshPortColors();
-            RefreshHideWiresButton();
         }
 
         public override void Dispose()
@@ -430,9 +289,9 @@ namespace Dynamo.ViewModels
             _node.WorkspaceViewModel.PropertyChanged -= Workspace_PropertyChanged;
         }
 
-        internal PortViewModel CreateProxyPortViewModel(PortModel portModel)
+        internal virtual PortViewModel CreateProxyPortViewModel(PortModel portModel)
         {
-            return new PortViewModel(_node, portModel);
+            throw new Exception("Don't do this");
         }
 
         /// <summary>
@@ -495,9 +354,6 @@ namespace Dynamo.ViewModels
                 case "ToolTipContent":
                     RaisePropertyChanged("ToolTipContent");
                     break;
-                case nameof(NodeViewModel.ZIndex):
-                    RefreshHideWiresButton();
-                    break;
             }
         }
 
@@ -516,9 +372,7 @@ namespace Dynamo.ViewModels
                     break;
                 case "IsConnected":
                     RaisePropertyChanged(nameof(IsConnected));
-                    RaisePropertyChanged(nameof(OutputPortBreakConnectionsButtonEnabled));
                     RefreshPortColors();
-                    RefreshHideWiresButton();
                     break;
                 case "IsEnabled":
                     RaisePropertyChanged("IsEnabled");
@@ -526,26 +380,8 @@ namespace Dynamo.ViewModels
                 case "Center":
                     RaisePropertyChanged("Center");
                     break;
-                case "DefaultValue":
-                    RaisePropertyChanged("DefaultValue");
-                    break;
-                case "UsingDefaultValue":
-                    RaisePropertyChanged("UsingDefaultValue");
-                    RefreshPortColors();
-                    break;
                 case "MarginThickness":
                     RaisePropertyChanged("MarginThickness");
-                    break;
-                case "UseLevels":
-                    RaisePropertyChanged("UseLevels");
-                    RaisePropertyChanged(nameof(UseLevelsMenuColor));
-                    break;
-                case "Level":
-                    RaisePropertyChanged("Level");
-                    break;
-                case "KeepListStructure":
-                    RaisePropertyChanged("ShouldKeepListStructure");
-                    RefreshPortColors();
                     break;
             }
         }
@@ -553,146 +389,43 @@ namespace Dynamo.ViewModels
         /// <summary>
         /// UseLevels command
         /// </summary>
+        [Obsolete("This method will be removed in Dynamo 3.0 - please use the InPortViewModel")]
         public DelegateCommand UseLevelsCommand
         {
             get
             {
                 if (_useLevelsCommand == null)
                 {
-                    _useLevelsCommand = new DelegateCommand(UseLevel, p => true);
+                    _useLevelsCommand = new DelegateCommand(null, p => true);
                 }
                 return _useLevelsCommand;
             }
         }
 
-        private void UseLevel(object parameter)
-        {
-            var useLevel = (bool)parameter;
-            var command = new DynamoModel.UpdateModelValueCommand(
-                Guid.Empty, _node.NodeLogic.GUID, "UseLevels", string.Format("{0}:{1}", _port.Index, useLevel));
-            
-            _node.WorkspaceViewModel.DynamoViewModel.ExecuteCommand(command);
-            RaisePropertyChanged(nameof(UseLevelSpinnerVisible));
-        }
-
         /// <summary>
         /// ShouldKeepListStructure command
         /// </summary>
+        [Obsolete("This method will be removed in Dynamo 3.0 - please use the InPortViewModel")]
         public DelegateCommand KeepListStructureCommand
         {
             get
             {
                 if (_keepListStructureCommand == null)
                 {
-                    _keepListStructureCommand = new DelegateCommand(KeepListStructure, p => true);
+                    _keepListStructureCommand = new DelegateCommand(null, p => true);
                 }
                 return _keepListStructureCommand;
             }
         }
 
-        /// <summary>
-        /// Used by the 'Break Connection' button in the node output context menu.
-        /// Removes any current connections this port has.
-        /// </summary>
-        public DelegateCommand BreakConnectionsCommand
-        {
-            get
-            {
-                if (_breakConnectionsCommand == null)
-                {
-                    _breakConnectionsCommand = new DelegateCommand(BreakConnections);
-                }
-                return _breakConnectionsCommand;
-            }
-        }
-
-        /// <summary>
-        /// Used by the 'Break Connection' button in the node output context menu.
-        /// Removes any current connections this port has.
-        /// </summary>
-        public DelegateCommand HideConnectionsCommand
-        {
-            get
-            {
-                if (_hideConnectionsCommand == null)
-                {
-                    _hideConnectionsCommand = new DelegateCommand(HideConnections);
-                }
-                return _hideConnectionsCommand;
-            }
-        }
-
-        private void KeepListStructure(object parameter)
-        {
-            bool keepListStructure = (bool)parameter;
-            var command = new DynamoModel.UpdateModelValueCommand(
-                Guid.Empty, _node.NodeLogic.GUID, "KeepListStructure", string.Format("{0}:{1}", _port.Index, keepListStructure));
-            
-            _node.WorkspaceViewModel.DynamoViewModel.ExecuteCommand(command);
-        }
-
+        [Obsolete("This method will be removed in Dynamo 3.0 - please use the InPortViewModel")]
         private void ChangeLevel(int level)
         {
             var command = new DynamoModel.UpdateModelValueCommand(
-                            Guid.Empty, _node.NodeLogic.GUID, "ChangeLevel", string.Format("{0}:{1}", _port.Index, level));
+                Guid.Empty, _node.NodeLogic.GUID, "ChangeLevel", string.Format("{0}:{1}", _port.Index, level));
 
             _node.WorkspaceViewModel.DynamoViewModel.ExecuteCommand(command);
         }
-
-        /// <summary>
-        /// Used by the 'Break Connection' button in the node output context menu.
-        /// Removes any current connections this port has.
-        /// </summary>
-        /// <param name="parameter"></param>
-        private void BreakConnections(object parameter)
-        {
-            for (int i = _port.Connectors.Count - 1; i >= 0; i--)
-            {
-                // Attempting to get the relevant ConnectorViewModel via matching GUID
-                ConnectorViewModel connectorViewModel = _node.WorkspaceViewModel.Connectors
-                    .FirstOrDefault(x => x.ConnectorModel.GUID == _port.Connectors[i].GUID);
-
-                if (connectorViewModel == null) continue;
-
-                connectorViewModel.BreakConnectionCommand.Execute(null);
-            }
-        }
-
-        /// <summary>
-        /// Used by the 'Hide Wires' button in the node output context menu.
-        /// Turns of the visibility of any connections this port has.
-        /// </summary>
-        /// <param name="parameter"></param>
-        private void HideConnections(object parameter)
-        {
-            for (int i = _port.Connectors.Count - 1; i >= 0; i--)
-            {
-                // Attempting to get the relevant ConnectorViewModel via matching GUID
-                ConnectorViewModel connectorViewModel = _node.WorkspaceViewModel.Connectors
-                    .FirstOrDefault(x => x.ConnectorModel.GUID == _port.Connectors[i].GUID);
-
-                if (connectorViewModel == null) continue;
-
-                connectorViewModel.HideConnectorCommand.Execute(!SetConnectorsVisibility);
-            }
-            RefreshHideWiresButton();
-        }
-        /// <summary>
-        /// Returns true if they are hidden.
-        /// </summary>
-        /// <returns></returns>
-        private bool CheckIfConnectorsAreHidden()
-        {
-            if (_port.Connectors.Count < 1 || _node.WorkspaceViewModel.Connectors.Count < 1) return false;
-
-            // Attempting to get a relevant ConnectorViewModel via matching NodeModel GUID
-            ConnectorViewModel connectorViewModel = _node.WorkspaceViewModel.Connectors
-                .FirstOrDefault(x => x.Nodevm.NodeModel.GUID == _port.Owner.GUID);
-
-            if (connectorViewModel == null) return false;
-            return connectorViewModel.IsCollapsed;
-        }
-
 
         private void Connect(object parameter)
         {
@@ -701,7 +434,7 @@ namespace Dynamo.ViewModels
             workspaceViewModel.HandlePortClicked(this);
         }
 
-        private bool CanConnect(object parameter)
+        protected bool CanConnect(object parameter)
         {
             return true;
         }
@@ -778,41 +511,10 @@ namespace Dynamo.ViewModels
         /// <summary>
         /// Handles the logic for updating the PortBackgroundColor and PortBackgroundBrushColor
         /// </summary>
-        private void RefreshPortColors()
+        protected virtual void RefreshPortColors()
         {
-            // The visual appearance of ports can be affected by many different logical states
-            // Inputs have more display styles than outputs
-            if (_port.PortType == PortType.Input)
-            {
-                // Special case for keeping list structure visual appearance
-                if (_port.UseLevels && _port.KeepListStructure && _port.IsConnected)
-                {
-                    PortValueMarkerColor = PortValueMarkerBlue;
-                    PortBackgroundColor = PortBackgroundColorKeepListStructure;
-                    PortBorderBrushColor = PortBorderBrushColorKeepListStructure;
-                }
-                // Port has a default value, shows blue marker
-                else if (UsingDefaultValue && DefaultValueEnabled)
-                {
-                    PortValueMarkerColor = PortValueMarkerBlue;
-                    PortBackgroundColor = PortBackgroundColorDefault;
-                    PortBorderBrushColor = PortBorderBrushColorDefault;
-                }
-                // Port isn't connected and has no default value (or isn't using it)
-                else
-                {
-                    PortValueMarkerColor = !_port.IsConnected ? PortValueMarkerRed : PortValueMarkerBlue;
-                    PortBackgroundColor = PortBackgroundColorDefault;
-                    PortBorderBrushColor = PortBorderBrushColorDefault;
-                }
-            }
-            // It's an output port, which either displays a connected style or a disconnected style
-            else
-            {
-                PortBackgroundColor = PortBackgroundColorDefault;
-                PortBorderBrushColor = PortBorderBrushColorDefault;
-            }
-            RaisePropertyChanged(nameof(UseLevelsMenuColor));
+            PortBackgroundColor = PortBackgroundColorDefault;
+            PortBorderBrushColor = PortBorderBrushColorDefault;
         }
 
         /// <summary>
